newEdit

+++ controllers/categoryController.js
@@ -1,6 +1,7 @@
 const mongoose = require("mongoose");
 const Category = require("./../models/category");
 const express = require("express");
+const { validationResult } = require("express-validator");
 
 exports.show_category = (request, response, next) => {
   Category.find({})
@@ -12,24 +13,42 @@ exports.show_category = (request, response, next) => {
     });
 };
 
-
 exports.delete_category = (request, response, next) => {
-  if(request.role == "admin"){
-    Category.findByIdAndDelete({ _id: request.body.id })
-    .then((data) => {
-      response.status(201).json({ message: "deleted", data });
-    })
-    .catch((error) => {
-      next(error);
+  let errors = validationResult(request);
+  if (!errors.isEmpty()) {
+    let error = new Error();
+    error.status = 422;
+    error.message = errors.array().reduce((current, object) => {
+      current + object.msg + " ", "";
     });
+    throw error;
   }
-  else{
-    response.status(404).json({msg:"You should be admin to delete category"})
+  if (request.role == "admin") {
+    Category.findByIdAndDelete({ _id: request.body.id })
+      .then((data) => {
+        response.status(201).json({ message: "deleted", data });
+      })
+      .catch((error) => {
+        next(error);
+      });
+  } else {
+    response
+      .status(404)
+      .json({ msg: "You should be admin to delete category" });
   }
 };
 
 exports.update_category = (request, response, next) => {
-  if(request.role == "admin"){
+  let errors = validationResult(request);
+  if (!errors.isEmpty()) {
+    let error = new Error();
+    error.status = 422;
+    error.message = errors.array().reduce((current, object) => {
+      current + object.msg + " ", "";
+    });
+    throw error;
+  }
+  if (request.role == "admin") {
     Category.findByIdAndUpdate(
       { _id: request.body.id },
       {
@@ -38,21 +57,30 @@ exports.update_category = (request, response, next) => {
           description: request.body.description,
           bannerImage: request.body.bannerImage,
         },
-      })
+      }
+    )
       .then((data) => {
         response.status(201).json({ message: " updated", data });
       })
       .catch((error) => {
         next(error);
       });
+  } else {
+    response.status(404).json({ msg: "You should be admin to edit category" });
   }
-  else{
-    response.status(404).json({msg:"You should be admin to edit category"})
-  }
-  };
+};
 
 exports.add_category = (request, response, next) => {
-  if(request.role == "admin"){
+  let errors = validationResult(request);
+  if (!errors.isEmpty()) {
+    let error = new Error();
+    error.status = 422;
+    error.message = errors.array().reduce((current, object) => {
+      current + object.msg + " ", "";
+    });
+    throw error;
+  }
+  if (request.role == "admin") {
     let object = new Category({
       // _id: auto_id,
       name: request.body.name,
@@ -65,10 +93,9 @@ exports.add_category = (request, response, next) => {
         response.status(201).json({ message: "added", data });
       })
       .catch((error) => next(error));
+  } else {
+    response.status(404).json({ msg: "You should be admin to add category" });
   }
-  else{
-    response.status(404).json({msg:"You should be admin to add category"})
-  }
-  
+
   //   });
 };
